+++ packages/react-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pantheon-systems/pcc-react-sdk
 
+## 3.3.2
+
+### Patch Changes
+
+- @pantheon-systems/pcc-sdk-core@3.3.2
+
 ## 3.3.1
 
 ### Patch Changes